review changes and newConnection method modified.
