+++ go/test/endtoend/backup/transform/mysqlctld/backup_transform_mysqlctld_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package transformmysqlctld
+package mysqlctld
 
 import (
 	"testing"
