+++ go/test/endtoend/mysqlctl/mysqlctl_test.go
@@ -58,7 +58,7 @@ func TestMain(m *testing.M) {
 
 		initCluster([]string{"0"}, 2)
 
-		// Collect table paths and ports
+		// Collect tablet paths and ports
 		tablets := clusterInstance.Keyspaces[0].Shards[0].Vttablets
 		for _, tablet := range tablets {
 			if tablet.Type == "master" {
