+++ go/test/endtoend/mysqlctld/mysqlctld_test.go
@@ -59,7 +59,7 @@ func TestMain(m *testing.M) {
 			return 1
 		}
 
-		// Collect table paths and ports
+		// Collect tablet paths and ports
 		tablets := clusterInstance.Keyspaces[0].Shards[0].Vttablets
 		for _, tablet := range tablets {
 			if tablet.Type == "master" {
