+++ test/config.json
@@ -192,7 +192,7 @@
 		},
 		"backup_transform_mysqlctld": {
 			"File": "backup_transform_mysqlctld_test.go",
-			"Args": ["vitess.io/vitess/go/test/endtoend/backup/transformmysqlctld"],
+			"Args": ["vitess.io/vitess/go/test/endtoend/backup/transform/mysqlctld"],
 			"Command": [],
 			"Manual": false,
 			"Shard": 21,
